requirements.txt
astroid==2.3.3
attrs==19.3.0
Automat==0.8.0
autopep8==1.5
cffi==1.13.2
constantly==15.1.0
coverage==4.5.1
cryptography==2.8
cssselect==1.1.0
Django==2.2
django-cors-headers==3.2.1
django-nose==1.4.6
djangorestframework==3.10.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
nose==1.3.7
parsel==1.5.2
promise==2.3
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pytz==2019.3
queuelib==1.5.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
typed-ast==1.4.1
w3lib==1.21.0
wrapt==1.11.2
zope.interface==4.7.1
